+++ components/Navbar/styledNavbar.js
@@ -8,14 +8,20 @@
 
 import styled from 'react-emotion';
 import { display } from 'styled-system';
+import theming from 'styled-theming';
 
 import theme from '../../style/theme';
 import media from '../../style/media';
 import { navContainerFragment } from '../Container';
 
+const backgroundColor = theming('category', {
+  library: theme.colors.blues.dark,
+  classroom: theme.colors.pinks.dark
+});
+
 export const Bar = styled('div')`
   position: relative;
-  background: ${theme.colors.blues.dark};
+  background: ${backgroundColor};
 `;
 
 export const NavItem = styled('div')`
